Fix validation tests
