+++ docs/src/info.md
@@ -5,8 +5,12 @@ CurrentModule = ZipStreams
 ```
 
 You can print information about a ZIP archive or a file within a ZIP archive to the terminal
-(or any `IO` object) with the `print_info` method. The format of the information displayed for
-each file is similar to the short format produced by [`ZipInfo`](https://linux.die.net/man/1/zipinfo).
+(or any `IO` object) with the three-argument `show` method, specifically
+`show(::IO, ::MIME"text/plain", file_or_archive)`. Note that this method is automatically called
+by the REPL to display such an object if it is the return value of the most recent statement. The
+format of the information displayed for each file is similar to the short format produced by
+[`ZipInfo`](https://linux.die.net/man/1/zipinfo).
+
 In general, the output has the following format:
 
 ```
@@ -29,7 +33,7 @@ so far is printed in archive order, along with status information about how much
 read from the archive so far, whether or not the EOF has been reached, and statistics about
 the number and size of the entries.
 
-The first call to `print_info` in this example reports nothing has been read yet:
+The first call to `show` in this example reports nothing has been read yet:
 
 ```@meta
 DocTestFilters = [r"\d{2}-[A-Z][a-z]{2}-\d{2} \d{2}:\d{2}:\d{2}"]
@@ -38,31 +42,26 @@ DocTestFilters = [r"\d{2}-[A-Z][a-z]{2}-\d{2} \d{2}:\d{2}:\d{2}"]
 ```jldoctest info1
 using ZipStreams
 
-buffer = IOBuffer()
+buffer = IOBuffer();
 zipsink(buffer) do sink
-    write_file(sink, "hello.txt", "Hello, Julia!")
-    write_file(sink, "subdir/goodbye.txt", "Goodbye, Julia!"; compression=:store, make_path=true)
+    write_file(sink, "hello.txt", "Hello, Julia!");
+    write_file(sink, "subdir/goodbye.txt", "Goodbye, Julia!"; compression=:store, make_path=true);
 end
 
-seekstart(buffer)
-zipsource(buffer) do source
-    print_info(source)
-end
+seekstart(buffer);
+source = zipsource(buffer)
 
 # output
 
 ZIP archive source stream data after reading 0 B, number of entries: 0
 ```
 
-After reading all of the data in the archive using `vialidate`, a call to `print_info` reports
+After reading all of the data in the archive using `vialidate`, the next thing that is shown is
 information about the entities read:
 
 ```jldoctest info1
-seekstart(buffer)
-zipsource(buffer) do source
-    validate(source)
-    print_info(source)
-end
+validate(source);
+source
 
 # output
 
@@ -79,5 +78,5 @@ DocTestFilters = nothing
 
 ## API
 ```@docs
-print_info
+show
 ```
