+++ src/convenience.jl
@@ -87,14 +87,14 @@ function unzip_files(archive_filename::AbstractString, files::AbstractVector{<:A
     files_to_extract = Set(files)
     zipsource(archive_filename) do source
         for file in source
-            if !isempty(files_to_extract) && file_info(file).name ∉ files_to_extract
+            if !isempty(files_to_extract) && info(file).name ∉ files_to_extract
                 continue
             end
-            dirs = split(file_info(file).name, ZIP_PATH_DELIMITER)[1:end-1]
+            dirs = split(info(file).name, ZIP_PATH_DELIMITER)[1:end-1]
             if !isempty(dirs)
                 mkpath(joinpath(output_path, dirs...))
             end
-            open(joinpath(output_path, file_info(file).name), "w") do io
+            open(joinpath(output_path, info(file).name), "w") do io
                 write(io, file)
             end
         end
